requirements.txt
azure-core==1.10.0
azure-storage-blob==12.7.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
configparser==5.0.1
cryptography==3.3.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
pycparser==2.20
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
